Project1/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
torchvision==0.15.2+cpu
transformers
boto3
pillow